Canonize Go KMS samples
