+++ kms/get_key.go
@@ -1,0 +1,59 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package kms contains samples for asymmetric keys feature of Cloud Key Management Service
+// https://cloud.google.com/kms/
+package kms
+
+// [START kms_get_asymmetric_public]
+
+import (
+	"context"
+	"crypto/x509"
+	"encoding/pem"
+	"fmt"
+
+	cloudkms "cloud.google.com/go/kms/apiv1"
+	kmspb "google.golang.org/genproto/googleapis/cloud/kms/v1"
+)
+
+// getAsymmetricPublicKey retrieves the public key from a saved asymmetric key pair on KMS.
+// example keyName: "projects/PROJECT_ID/locations/global/keyRings/RING_ID/cryptoKeys/KEY_ID/cryptoKeyVersions/1"
+func getAsymmetricPublicKey(keyName string) (interface{}, error) {
+	ctx := context.Background()
+	client, err := cloudkms.NewKeyManagementClient(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	// Build the request.
+	req := &kmspb.GetPublicKeyRequest{
+		Name: keyName,
+	}
+	// Call the API.
+	response, err := client.GetPublicKey(ctx, req)
+	if err != nil {
+		return nil, fmt.Errorf("failed to fetch public key: %+v", err)
+	}
+	// Parse the key.
+	keyBytes := []byte(response.Pem)
+	block, _ := pem.Decode(keyBytes)
+	publicKey, err := x509.ParsePKIXPublicKey(block.Bytes)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse public key: %+v", err)
+	}
+	return publicKey, nil
+}
+
+// [END kms_get_asymmetric_public]
